refactor: RoundDto 패키지 분리 및 ConnectionManger 클래스 이름 변경
