+++ src/main/resources/static/index.html
@@ -97,7 +97,7 @@
 <div>
     <button onclick="ajaxScore()">점수확인 버튼</button>
 </div>
-<div id="board"><span id="score">asd</span></div>
+<div id="board"><span id="score"></span></div>
 </body>
 <script src="app.js"></script>
 </html>
